+++ Jellyfin.Plugin.JellyfinHelper.Tests/Services/Statistics/MediaStatisticsServiceTrashPathResolutionTests.cs
@@ -4,6 +4,7 @@ using Jellyfin.Plugin.JellyfinHelper.Services.Cleanup;
 using Jellyfin.Plugin.JellyfinHelper.Services.PluginLog;
 using Jellyfin.Plugin.JellyfinHelper.Services.Statistics;
 using Jellyfin.Plugin.JellyfinHelper.Tests.TestFixtures;
+using MediaBrowser.Common.Configuration;
 using MediaBrowser.Controller.Entities;
 using MediaBrowser.Controller.Library;
 using MediaBrowser.Model.Entities;
@@ -40,7 +41,8 @@ public sealed class MediaStatisticsServiceTrashPathResolutionTests
             _fileSystemMock.Object,
             _pluginLogMock.Object,
             TestMockFactory.CreateLogger<MediaStatisticsService>().Object,
-            configHelper);
+            configHelper,
+            TestMockFactory.CreateAppPaths().Object);
 
     [Fact]
     public void CalculateStatistics_TrashPathResolutionThrowsArgumentException_LogsWarningAndCompletesScan()